+++ src/NeuralNetworks/Layers/DiffusionConvLayer.cs
@@ -1665,7 +1665,7 @@ public partial class DiffusionConvLayer<T> : LayerBase<T>
     /// <summary>
     /// Gets the total number of trainable parameters.
     /// </summary>
-    public override int ParameterCount => _weights.Length + _biases.Length + DiffusionTimes.Length;
+    public override long ParameterCount => _weights.Length + _biases.Length + DiffusionTimes.Length;
 
     /// <summary>
     /// Creates a deep copy of this layer.
