+++ src/NeuralNetworks/Layers/MeshEdgeConvLayer.cs
@@ -699,7 +699,7 @@ public partial class MeshEdgeConvLayer<T> : LayerBase<T>
     /// <summary>
     /// Gets the total number of trainable parameters.
     /// </summary>
-    public override int ParameterCount => _weights.Length + _biases.Length;
+    public override long ParameterCount => _weights.Length + _biases.Length;
 
     /// <summary>
     /// Creates a deep copy of the layer.
