+++ src/NeuralNetworks/Layers/RRDBLayer.cs
@@ -115,7 +115,7 @@ public class RRDBLayer<T> : LayerBase<T>
     public int GrowthChannels => _growthChannels;
 
     /// <inheritdoc />
-    public override int ParameterCount => GetParameters().Length;
+    public override long ParameterCount => GetParameters().Length;
     public override bool SupportsTraining => true;
 
     /// <summary>
