+++ src/NeuralNetworks/Layers/SSM/GatedDeltaNetLayer.cs
@@ -177,7 +177,7 @@ public partial class GatedDeltaNetLayer<T> : LayerBase<T>
     /// <summary>
     /// Gets the total number of trainable parameters.
     /// </summary>
-    public override int ParameterCount =>
+    public override long ParameterCount =>
         _convWeights.Length + _convBias.Length +
         _queryWeights.Length + _keyWeights.Length + _valueWeights.Length +
         _betaWeights.Length + _betaBias.Length +
@@ -556,7 +556,7 @@ public partial class GatedDeltaNetLayer<T> : LayerBase<T>
     /// <inheritdoc />
     public override Vector<T> GetParameters()
     {
-        var parameters = new Vector<T>(ParameterCount);
+        var parameters = new Vector<T>((int)ParameterCount);
         int index = 0;
         foreach (var tensor in GetAllTensors())
             for (int i = 0; i < tensor.Length; i++)
@@ -587,7 +587,7 @@ public partial class GatedDeltaNetLayer<T> : LayerBase<T>
 
     public override Vector<T> GetParameterGradients()
     {
-        if (_convWeightsGradient == null) return new Vector<T>(ParameterCount);
+        if (_convWeightsGradient == null) return new Vector<T>((int)ParameterCount);
         return Vector<T>.Concatenate(
             new Vector<T>(_convWeightsGradient?.ToArray() ?? Array.Empty<T>()),
             new Vector<T>(_convBiasGradient?.ToArray() ?? Array.Empty<T>()),
