+++ src/NeuralNetworks/Layers/SSM/PaTHAttentionLayer.cs
@@ -145,7 +145,7 @@ public partial class PaTHAttentionLayer<T> : LayerBase<T>
     public int HeadDimension => _headDimension;
 
     /// <inheritdoc />
-    public override int ParameterCount =>
+    public override long ParameterCount =>
         _queryWeights.Length + _keyWeights.Length + _valueWeights.Length +
         _householderVectors.Length +
         _outputGateWeights.Length + _outputGateBias.Length +
@@ -550,7 +550,7 @@ public partial class PaTHAttentionLayer<T> : LayerBase<T>
     /// <inheritdoc />
     public override Vector<T> GetParameters()
     {
-        var parameters = new Vector<T>(ParameterCount);
+        var parameters = new Vector<T>((int)ParameterCount);
         int index = 0;
         foreach (var tensor in GetAllTensors())
             for (int i = 0; i < tensor.Length; i++)
@@ -579,7 +579,7 @@ public partial class PaTHAttentionLayer<T> : LayerBase<T>
 
     public override Vector<T> GetParameterGradients()
     {
-        if (_queryWeightsGradient == null) return new Vector<T>(ParameterCount);
+        if (_queryWeightsGradient == null) return new Vector<T>((int)ParameterCount);
         return Vector<T>.Concatenate(
             new Vector<T>(_queryWeightsGradient?.ToArray() ?? Array.Empty<T>()),
             new Vector<T>(_keyWeightsGradient?.ToArray() ?? Array.Empty<T>()),
