+++ src/NeuralNetworks/Layers/SSM/TTTLayer.cs
@@ -176,7 +176,7 @@ public partial class TTTLayer<T> : LayerBase<T>
     /// <summary>
     /// Gets the total number of trainable parameters.
     /// </summary>
-    public override int ParameterCount =>
+    public override long ParameterCount =>
         _queryWeights.Length + _queryBias.Length +
         _keyWeights.Length + _keyBias.Length +
         _valueWeights.Length + _valueBias.Length +
@@ -662,7 +662,7 @@ public partial class TTTLayer<T> : LayerBase<T>
     /// <inheritdoc />
     public override Vector<T> GetParameters()
     {
-        var parameters = new Vector<T>(ParameterCount);
+        var parameters = new Vector<T>((int)ParameterCount);
         int index = 0;
         foreach (var tensor in GetAllTensors())
             for (int i = 0; i < tensor.Length; i++)
@@ -694,7 +694,7 @@ public partial class TTTLayer<T> : LayerBase<T>
 
     public override Vector<T> GetParameterGradients()
     {
-        if (_queryWeightsGradient == null) return new Vector<T>(ParameterCount);
+        if (_queryWeightsGradient == null) return new Vector<T>((int)ParameterCount);
         return Vector<T>.Concatenate(
             new Vector<T>(_queryWeightsGradient?.ToArray() ?? Array.Empty<T>()),
             new Vector<T>(_queryBiasGradient?.ToArray() ?? Array.Empty<T>()),
