+++ src/NeuralNetworks/Layers/SeparableConvolutionalLayer.cs
@@ -334,13 +334,13 @@ public partial class SeparableConvolutionalLayer<T> : LayerBase<T>
     /// through a process called backpropagation, which adjusts its internal values based on errors it makes.
     /// </para>
     /// </remarks>
-    public override int ParameterCount => _depthwiseKernels.Length + _pointwiseKernels.Length + _biases.Length;
+    public override long ParameterCount => _depthwiseKernels.Length + _pointwiseKernels.Length + _biases.Length;
     public override bool SupportsTraining => true;
 
     public override Vector<T> GetParameterGradients()
     {
         if (_depthwiseKernelsGradient == null || _pointwiseKernelsGradient == null || _biasesGradient == null)
-            return new Vector<T>(ParameterCount);
+            return new Vector<T>((int)ParameterCount);
         return Vector<T>.Concatenate(
             Vector<T>.Concatenate(new Vector<T>(_depthwiseKernelsGradient.ToArray()), new Vector<T>(_pointwiseKernelsGradient.ToArray())),
             new Vector<T>(_biasesGradient.ToArray()));
