+++ src/NeuralNetworks/Layers/SpatialTransformerLayer.cs
@@ -359,7 +359,7 @@ public partial class SpatialTransformerLayer<T> : LayerBase<T>, IAuxiliaryLossLa
     /// The spatial transformer will gradually learn the best transformations for the task at hand.
     /// </para>
     /// </remarks>
-    public override int ParameterCount => GetParameters().Length;
+    public override long ParameterCount => GetParameters().Length;
     public override bool SupportsTraining => true;
 
     /// <summary>
